Getting all Chicago PurpleAir sensors
Wrote a script to download, using the (new) PurpleAir API, basic info
for all the PurpleAir sensors in the Chicago area. Here the Chicago area
is defined to be the bounding box containing the Chicago city
boundaires. The boundaries are from the Chicago Data Portal and stored
in geojson format. The resulting info for the sensors is stored as
chicago_sensors.csv.